Refactor mapper.
