+++ atlasdb-client/src/main/java/com/palantir/atlasdb/cache/LongEntryMapper.java
@@ -23,7 +23,7 @@ import com.palantir.logsafe.Preconditions;
 
 import okio.ByteString;
 
-public class LongEntryMapper implements DefaultOffHeapCache.EntryMapper<Long, Long> {
+public class LongEntryMapper implements EntryMapper<Long, Long> {
     @Override
     public ByteString serializeKey(Long key) {
         Preconditions.checkNotNull(key, "Key should not be null");
